+++ headers/RadarKit/RKSweep.h
@@ -60,7 +60,7 @@ struct rk_sweep_engine {
     char                             statusBuffer[RKBufferSSlotCount][RKMaximumStringLength];
     uint32_t                         statusBufferIndex;
     RKEngineState                    state;
-    uint32_t                         tic;
+    uint64_t                         tic;
     float                            lag;
     uint32_t                         almostFull;
     size_t                           memoryUsage;
